# src/resources/openapi/swagger_apartment_management.yaml
openapi: 3.0.3
info:
  title: Swagger Apartment Management System - OpenAPI 3.0
  description: |-
    This is a sample Apartment Management System Server based on the OpenAPI 3.0 specification.
    Some useful links:
    - [The Apartment Management System repository](https://github.com/swagger-api/swagger-petstore)
  termsOfService: http://swagger.io/terms/
  contact:
    email: 283531@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://pwrapartmentmanagement/api/v3
...
paths:
...
  /apartment/user:
    post:
      tags:
        - user
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
    get:
      tags:
        - user
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
  /apartment/user/login:
    post:
      tags:
        - user
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
  /apartment/user/{userID}:
    get:
      tags:
        - user
      summary: Get a specific user by ID
      operationId: getUserById
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
  /apartment/user/{userID}/delete:
    delete:
      tags:
        - user
      summary: Delete a user by ID
      operationId: deleteUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
  /apartment/user/{userID}/update:
    post:
      tags:
        - user
      summary: Update an existing user
      operationId: updateUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid property ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid property ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The property could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The property could not be found
...
components:
  schemas:
    SuccessApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          default: 201
        message:
          type: string
          default: "Success."
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    UserDTO:
      type: object
      properties:
        name:
          type: string
          example: "John"
        surname:
          type: string
          example: "Doe"
        email:
          type: string
          example: "john.doe@example.com"
        phoneNumber:
          type: string
          example: "+1234567890"
        password:
          type: string
          example: "securePassword123"
        flatList:
          type: array
          items:
            type: string
          example: ["Flat101", "Flat102"]
        roleType:
          type: string
          example: "Tenant"
      required:
        - name
        - surname
        - email
        - phoneNumber
        - password
        - roleType
    UserLoginDTO:
      type: object
      properties:
        email:
          type: string
          example: "john.doe@example.com"
        password:
          type: string
          example: "securePassword123"
      required:
        - email
        - password
...